requirements.txt
torch
pandas
numpy
torchvision
pillow
matplotlib
runx